Added a stylesheet
Been messing around with css, I took away the photo I had for background as it just complicated things and I didnt like how it looked. I been also trying to make the text have a custom font but it hasnt been working yet. for the rest I been adjusting the text to be in the position I want it to be, its nearly there but I still need to move it more.